requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
docx2txt